requirements.txt
torch
transformers
pillow
whisper
pyaudio
langchain-community
langchain-ollama
blip-image-captioning
ollama
subprocess